requirements.txt
tensorboard==2.4.1
mcp[cli]>=1.9
requests>=2.31
python-dotenv>=1.0
